requirements.txt
langchain
python-dotenv==1.0.0
streamlit==1.22.0
tiktoken>=0.7.0
faiss-cpu==1.12.0
protobuf~=3.19.0
langchain_experimental
mysql-connector-python
pymysql
sentence-transformers
chromadb
langchain-google-genai
langchain-huggingface